Time: 83 ms (55.77%), Space: 68 MB (28.05%) - LeetHub

+++ 0053-maximum-subarray/0053-maximum-subarray.cpp
@@ -3,7 +3,7 @@ public:
     int maxSubArray(vector<int>& nums) {
         int i=0;
         int j =0;
-        long long sum = 0;
+        int sum = 0;
         int maxsum =INT_MIN;
 
         while(j<nums.size()){
